refactor: move is-node-from-template to dom
