+++ packages/@lwc/engine/src/framework/vm.ts
@@ -535,30 +535,6 @@ function getErrorBoundaryVM(vm: VM): VM | undefined {
     }
 }
 
-/**
- * EXPERIMENTAL: This function detects whether or not a Node is
- * controlled by a LWC template. This API is subject to
- * change or being removed.
- */
-export function isNodeFromTemplate(node: Node): boolean {
-    if (isFalse(node instanceof Node)) {
-        return false;
-    }
-    // TODO [#1250]: skipping the shadowRoot instances itself makes no sense, we need to revisit this with locker
-    if (node instanceof ShadowRoot) {
-        return false;
-    }
-    if (useSyntheticShadow) {
-        // TODO [#1252]: old behavior that is still used by some pieces of the platform, specifically, nodes inserted
-        // manually on places where `lwc:dom="manual"` directive is not used, will be considered global elements.
-        if (isUndefined((node as any).$shadowResolver$)) {
-            return false;
-        }
-    }
-    const root = node.getRootNode();
-    return root instanceof ShadowRoot;
-}
-
 // slow path routine
 // NOTE: we should probably more this routine to the synthetic shadow folder
 // and get the allocation to be cached by in the elm instead of in the VM
